agrego double a factor reliquia y defensa

+++ src/app/reliquia/AnilloElfo.java
@@ -2,7 +2,7 @@ package app.reliquia;
 
 public class AnilloElfo  extends AnilloPoder{
 
-    public AnilloElfo(String nombre, int factorDeAtaque, int factorDeDefensa, int energiaMagica) {
+    public AnilloElfo(String nombre, double factorDeAtaque, double factorDeDefensa, int energiaMagica) {
         super(nombre, factorDeAtaque, factorDeDefensa, energiaMagica);
     }
 }
