+++ src/app/reliquia/AnilloNarya.java
@@ -2,7 +2,7 @@ package app.reliquia;
 
 public class AnilloNarya extends AnilloElfo{
 
-    public AnilloNarya(String nombre, int factorDeAtaque, int factorDeDefensa, int energiaMagica) {
+    public AnilloNarya(String nombre, double factorDeAtaque, double factorDeDefensa, int energiaMagica) {
         super(nombre, factorDeAtaque, factorDeDefensa, energiaMagica);
     }
 }
